refactor(adapter): Rename tx to exec in create_user for clarity
The variable `tx` in the `create_user` method was an abbreviation for "transaction". However, the `ExecutorWrapper` type it represents can contain either a transaction or a regular database executor.

The name has been changed to `exec` to more accurately reflect that the argument can be either, improving the code's readability and maintainability.