refactor(repository): Add lifetime to UserSqliteSqlxRepositoryAdapter
The `Transaction` associated type in the `UserRepositoryPort` implementation was previously hardcoded with a `'static` lifetime. This is overly restrictive and does not accurately represent the lifetime of a database transaction.

This commit introduces a generic lifetime `'t` to the `UserSqliteSqlxRepositoryAdapter` and uses it for the `Transaction` type. This correctly ties the transaction's lifetime to the adapter, making the implementation more flexible and sound. `PhantomData` is used to associate the struct with the new lifetime parameter.